requirements.txt
selenium==4.9.1
moviepy==1.0.3
pandas==2.0.3
Pillow==10.0.0
numpy==1.24.3
webdriver-manager==3.8.6
urllib3==1.26.16